test: extract ExportTestSupport (4 artifacts/export mocks) from TestSupport (#553, #434)
Third slice of splitting the 1,048-line TestSupport.swift "junk drawer"
into domain files, one cohesive mock group per PR, names stable, green
between. Slices 434a (Recording) and 434b (Screenshot) shipped in #551
and #558.

Moves the four export/artifacts mocks — MockArtifactsService
(SessionArtifactsManaging), MockClipboardService (ClipboardWriting),
@MainActor MockDebugBundleExporter (DebugBundleExporting), and
@MainActor MockPrivacyDataExporter (PrivacyDataExporting) with its
nested ExportRequest struct — byte-identical from TestSupport.swift
lines 300-393 into a new ExportTestSupport.swift. Both @MainActor
annotations sit on the line immediately above the corresponding class
declaration, preserved verbatim.

The new file's imports are the minimum required: Foundation and
@testable import BugNarrator (no XCTest, AppKit, ScreenCaptureKit, or
UniformTypeIdentifiers — the mocks only store strings and use Foundation
for FileManager/URL/UUID/TimeInterval).

Every other mock (MockHotkeyManager, both actor issue-extraction/export
mocks, MockOperationalTelemetryRecorder, StubExportReceiptStore, etc.)
and AppStateHarness stay in TestSupport.swift. Pure test-support move:
no product code, no helper renames, no behavior change. The full
BugNarratorTests suite stays green at 662 tests (unchanged count from
the current main baseline); docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (4 findings folded
in: exact 300-393 range, byte-for-byte diff bound, minimum-imports list,
nested ExportRequest must move with parent).
